Partial first implementation of StemLite.
StemLite is going to be an experimental implementation of Stem.

The goal for StemLite is to be able to create a network of topologies
and simply load-balance function loaded from strings at runtime. A
next version, based on the same (more matured) contract, will be
implemented to take .jar (or exec binaries, not sure) into account.